pyproject.toml
[project]
name = "mcp-gauge"
version = "0.1.0"
description = "MCP server for E2E testing and evaluation of MCP servers, enabling autonomous development by coding agents"
requires-python = ">=3.12"
license = "MIT"
keywords = ["mcp", "testing", "llm", "autonomous-development"]
dependencies = [
    "mcp>=1.0,<2.0",
    "anthropic>=0.30,<1.0",
    "aiosqlite>=0.20",
    "pydantic>=2.0,<3.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "mypy>=1.0",
    "ruff>=0.8.0",
    "pre-commit>=4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_gauge"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/mcp_gauge"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    ".steering/*",
    "src/mcp_gauge/__main__.py",
    "src/mcp_gauge/server.py",
]

[tool.coverage.report]
fail_under = 80
show_missing = true

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[dependency-groups]
dev = [
    "types-pyyaml>=6.0.12.20250915",
]
